+++ src/layout/Main/Main.js
@@ -1,27 +1,11 @@
 import { Box } from "@mui/material";
 import PersistentDrawer from "./drawer/PersistentDrawer";
+import { mainBoxStyle } from "./useMainStyles";
 
 function Main({ isOpen, toggleDrawer }) {
   return (
-    <Box
-      sx={{
-        display: "flex",
-        position: "absolute",
-        height: "100%",
-        top: 64,
-        left: 0,
-        flexGrow: 1,
-        width: "100%",
-      }}
-    >
-      {/* <DraggableDrawer
-        isOpen={isOpen}
-        toggleDrawer={toggleDrawer}
-      /> */}
+    <Box sx={mainBoxStyle}>
       <PersistentDrawer isOpen={isOpen} toggleDrawer={toggleDrawer} />
-      {/* <MainContentArea>
-        <Outlet />
-      </MainContentArea> */}
     </Box>
   );
 }
